endterm/requirements.txt
Babel==2.7.0
colorama==0.4.1
Django==2.2.6
django-admin-list-filter-dropdown==1.0.2
django-appconf==1.0.3
django-autocomplete-light==3.4.1
django-cors-headers==3.0.2
django-dramatiq==0.7.1
django-filter==2.2.0
django-imagekit==4.0.2
django-object-actions==1.1.0
django-phonenumber-field==3.0.1
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
dramatiq==1.6.1
httpie==1.0.2
Markdown==3.1.1
model-mommy==1.6.0
pilkit==2.0
Pillow==6.2.1
prometheus-client==0.7.1
Pygments==2.4.2
PyJWT==1.7.1
pytz==2019.3
pywin32==224
sqlparse==0.3.0
virtualenv==16.6.2
xlwt==1.3.0
